refactor(session): decouple ChatMessage from tui feature; dedupe Running→Done collapse
Addresses Task 2+3 review feedback: move ChatMessage/ToolStatus into
src/session/message.rs (ungated), move SessionId::short() to id.rs,
consolidate duration_ms serde adaptor, drop dead Running→Done branch in
ChatMessage Serialize (ToolStatus::Serialize already handles it).
