+++ src/session/mod.rs
@@ -4,7 +4,9 @@
 //! can be resumed later. See `docs/superpowers/specs/2026-04-18-tui-sessions-and-polish-design.md`.
 
 pub mod id;
+pub mod message;
 pub mod types;
 
 pub use id::SessionId;
+pub use message::{ChatMessage, ToolStatus};
 pub use types::{MessageCounts, Session, SessionMeta, StoredMessage};
